ext2: Support concurrent connections
Instead of serializing connections (which in turn serializes requests
because it is unsafe to open more than one handle into the filesystem
at the same time), use the recently-added APIs to open the filesystem
during .after_fork, then use the fact that ext2 code then supports
parallel access to files within the single system handle.  Still, even
if the underlying plugin accepts parallel actions, we serialize
because e2fsprogs is not re-entrant.