plugins/floppy: wrap source code at 80 chars
Wrap the source code at 80 characters.

In the create_virtual_floppy() function, we have a complex calculation
that is not really possible to squeeze under 80 characters in width,
without extracting part of it to a local variable. The problem is *naming*
the new variable, plus the *type* of the new variable.

By inspecting the rest of the function, I decided that the expression
being extracted, namely

  DIV_ROUND_UP ((data_size / CLUSTER_SIZE + 2) * 4, CLUSTER_SIZE)

should be called "fat_clusters". The rest of the function calculates the
same quantity, only in multiple steps; the result is then assigned to
"floppy->fat_clusters".

Regarding the type of the new local variable: "data_size" is "uint64_t",
and everything else in the expression is "int", so the result is
"uint64_t" as well, in practice.

Shorten the original assignment by using compound assignment, as well.

Tested with "make check"; "tests/test-floppy-size.sh" passes.

Cc: Nolan Leake <nolan@sigbus.net>
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2172516
Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20230413140346.908687-12-lersek@redhat.com>
[lersek@redhat.com: refresh last paragraph of commit message, according to
 new commit a7709a69ac08]