build: windows: Don't link with msvcrt
Attempting to link C programs with msvcrt on recent Fedora fails with:

*** Warning: linker path does not have real file for library -lmsvcrt.
*** I have the capability to make that library automatically link in when
*** you link to this library.  But I can only do this if you have a
*** shared version of the library, which you do not appear to have
*** because I did check the linker path looking for a file starting
*** with libmsvcrt and none of the candidates passed a file format test
*** using a file magic. Last file checked: /usr/x86_64-w64-mingw32/sys-root/mingw/lib/libmsvcrt.a

It seems some time between mingw-crt 5 and 10 (RHEL 8 and RHEL 9) the
import library (libmsvcrt.a) was changed in a way that libtool is no
longer able recognise it as an implib.

As Wine seems to work fine with *.exe and *.dll which don't link to
this library, drop it from the linker line to solve the above problem.

Commit a5112e9048 ("server: Port nbdkit to Windows.") introduced
linking with this library without any justification.