rust: Change 'open' method so it can return an error
The underlying plugin API allows the open method to return an error
(by returning NULL).  However this was not mirrored in the rust
bindings because you could only return a boxed object (equivalent to
returning a non-NULL object in C).

Modify the rust API so Result<> is required.  The API changes as
follows, requiring all plugins to change:

  - fn open(readonly: bool) -> Box<dyn Server> where Self: Sized;
  + fn open(readonly: bool) -> Result<Box<dyn Server>> where Self: Sized;

Trivially this can be done by surrounding the return value of the
existing open method with Ok(...)

Tested by applying the following patch to example.rs:

  --- a/plugins/rust/examples/ramdisk.rs
  +++ b/plugins/rust/examples/ramdisk.rs
  @@ -52,7 +52,8 @@ impl Server for RamDisk {
       }

       fn open(_readonly: bool) -> Result<Box<dyn Server>> {
  -        Ok(Box::<RamDisk>::default())
  +    //  Ok(Box::<RamDisk>::default())
  +        Err(Error::new(libc::EINVAL, "oops"))
       }

       fn read_at(&self, buf: &mut [u8], offset: u64) -> Result<()> {

and observing the behaviour when connecting to the plugin:

  nbdkit: ramdisk[1]: debug: ramdisk: open readonly=0 exportname="" tls=0
  nbdkit: ramdisk[1]: debug: ramdisk: default_export readonly=0 tls=0
  nbdkit: ramdisk[1]: error: oops

See: https://github.com/libguestfs/nbdkit/issues/21