allocators: sparse: Improve performance by using a read-write lock
We've long known that nbdkit-memory-plugin with the default sparse
array allocator suffers because a global mutex must be taken whenever
any read or write request is performed.  This commit aims to safely
improve performance by converting the mutex into a read-write lock.

The shared (read) lock still must be acquired for any access to the
sparse array.  However if you wish to make any changes to the L1/L2
directory, which includes allocating or deallocating pages, then you
must upgrade to the exclusive (write) lock.

Unfortunately POSIX read-write locks are not upgradable, and previous
attempts to work around this lead to many potential safety issues
and/or deadlocks.  For example if at the point where you want to
update the metadata, you unlock the shared lock and attempt to acquire
the exclusive lock, then other threads can jump in and modify the
metadata underneath you.  It's very difficult to work around this
limitation safely.  This patch takes a different approach: We attempt
write requests using the shared lock, but if we detect that we need to
update the metadata during an operation then we retry the entire
request from the start with the exclusive lock.  This should be safe,
albeit slower.

Either the test program (contrib/sparseloadtest.c), or fio, can be
used to test performance.  The test program is better because it
exercises page deallocation, which is not exercised by fio testing.

For unpatched nbdkit 1.34:

  - 4 threads, mix of 80% read, 10% write, 10% trim:

    READ: 77669.5 ops/s 5087519644.9 bytes/s
    WRITE: 9707.1 ops/s 635528046.7 bytes/s
    TRIM: 9712.0 ops/s 636737940.6 bytes/s
    TOTAL: 97088.6 ops/s 6359785632.1 bytes/s

  - 4 threads, mix of 20% read, 40% write, 40% trim:

    READ: 20135.1 ops/s 1318832390.2 bytes/s
    WRITE: 40293.6 ops/s 2640703848.7 bytes/s
    TRIM: 40240.7 ops/s 2636021905.5 bytes/s
    TOTAL: 100669.4 ops/s 6595558144.3 bytes/s

For patched nbdkit:

  - 4 threads, mix of 80% read, 10% write, 10% trim:

    READ: 75265.4 ops/s 4931584759.3 bytes/s
    WRITE: 9392.0 ops/s 614956850.5 bytes/s
    TRIM: 9426.0 ops/s 618151433.1 bytes/s
    TOTAL: 94083.4 ops/s 6164693042.9 bytes/s

    Note that performance is about 3% lower, possibly because
    read-write locks are less efficient than mutexes?

  - 4 threads, mix of 20% read, 40% write, 40% trim:

    READ: 23008.4 ops/s 1508451376.5 bytes/s
    WRITE: 46011.8 ops/s 3014531474.9 bytes/s
    TRIM: 46015.2 ops/s 3014830518.1 bytes/s
    TOTAL: 115035.3 ops/s 7537813369.6 bytes/s

    Performance is about 15% better.

With fio, unpatched nbdkit 1.34:

   READ: bw=1188MiB/s (1245MB/s), 1188MiB/s-1188MiB/s (1245MB/s-1245MB/s), io=69.6GiB (74.7GB), run=60001-60001msec
  WRITE: bw=1187MiB/s (1245MB/s), 1187MiB/s-1187MiB/s (1245MB/s-1245MB/s), io=69.6GiB (74.7GB), run=60001-60001msec

With fio, patched nbdkit:

   READ: bw=1309MiB/s (1372MB/s), 1309MiB/s-1309MiB/s (1372MB/s-1372MB/s), io=76.7GiB (82.3GB), run=60001-60001msec
  WRITE: bw=1308MiB/s (1372MB/s), 1308MiB/s-1308MiB/s (1372MB/s-1372MB/s), io=76.7GiB (82.3GB), run=60001-60001msec

fio test command:

  nbdkit -U - memory size=256M --run 'export uri; fio examples/nbd.fio'

with examples/nbd.fio from the fio git repository.

Reviewed-by: Eric Blake <eblake@redhat.com>