server: Add -D nbdkit.environ=1 to dump the environment
This is not secure so should not be used routinely.  Also we do not
attempt to filter environment variables, so even ones containing
multiple lines or special characters are all sent to nbdkit_debug.

The reason for adding this is to allow for debugging the new
nbd_set_socket_activation_name(3) API added in libnbd 1.16.

Reviewed-by: Laszlo Ersek <lersek@redhat.com>