server: debug: Fix error handling
Preserve errno even if the function fails.  We need to set
errno = err again on every exit path out of the function.

Check if close_memstream failed and go to the error path if so.  This
is most important on Windows where close_memstream is what actually
allocates the memory.  But even on Unix, close_memstream (which is an
alias for fclose) can return an error.

Move the error-handling code to the end of the function.

I added some clarifying comments:

 - above the 'fputs' explaining why exactly we're using memstream in
   the first place (ie soft atomicity)

 - at each assignment to errno

See: https://listman.redhat.com/archives/libguestfs/2023-May/031456.html
Reviewed-by: Eric Blake <eblake@redhat.com>
Thanks: Laszlo Ersek