rust: Bind nbdkit_is_tls as nbdkit::is_tls()
This returns the per-connection TLS flag which can be used to
determine if the current connection successfully negotiated TLS.  Call
it from the open() method.

I tested this by hand using the ramdisk example:

  $ ./nbdkit -fv --tls=require --tls-certificates=tests/pki \
        ./plugins/rust/target/release/examples/libramdisk.so \
        --run 'nbdinfo $uri'
  ...
  nbdkit: ramdisk[1]: debug: connection opened: readonly=false, tls=true
  ...
  protocol: newstyle-fixed with TLS, using structured packets
  export="":
	export-size: 104857600 (100M)
	content: data
	uri: nbds://localhost:10809/

Notice 'tls=true' in the debug message which comes from a call to
nbdkit::is_tls().  Notice also the protocol contains 'with TLS' which
indicates that libnbd negotiated a TLS connection.

Fixes: https://gitlab.com/nbdkit/nbdkit/-/issues/11