server: s/handle/cookie/ to match NBD spec
Externally, libnbd has been exposing the 64-bit opaque marker for each
NBD packet as the "cookie", because it was less confusing when
contrasted with 'struct nbd_handle *' holding all libnbd state.  It
also avoids confusion between the noun 'handle' as a way to identify a
packet and the verb 'handle' for reacting to things like signals.
Upstream NBD changed their spec to favor the name "cookie" based on
our recommendations[1], and so now we can get rid of our last uses of
the old name, keeping nbd-protocol.h in sync with libnbd.

[1] https://github.com/NetworkBlockDevice/nbd/commit/ca4392eb2b

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <b999b0fb-8233-fd4a-8e21-c4c19aff21bc@redhat.com>
Reviewed-by: Laszlo Ersek <lersek@redhat.com>