data: Allow more optimizations of string*N
We can replace string*N by a single string containing N copies.
Previously we only did this for N <= 4 and string length <= 512.

Change this so we test the total length of the final string, and allow
a more generous <= 4096 bytes for the final string.

For example this would allow "1234"*32 to be optimized.