tests/web-server.c: Ignore errors when reading the request
In some circumstances curl will shutdown the socket while we are
reading the request.  This seems to happen only under valgrind,
possibly because of the slower speed of the test program.  Anyway it
is quite safe to ignore an error reading the request and go back to
accepting a new request.

I also changed existing error messages so they can be distinguished.