ocaml: Fix thread registration for OCaml 5
OCaml 5 is strict about registering threads before calling OCaml heap
functions, and will abort the program with this error if you don't do
this correctly:

  Fatal error: no domain lock held

Fix this as explained in the comment.

Note (as it's not explained well in the documentation): Threads
created from C are placed in OCaml thread domain 0.  In order to add
them to this domain, the main program must not hold on to the runtime
system lock (for domain 0, because that's what caml_startup gives
you).  For this to work we must only hold this lock in the main
program briefly around calls to OCaml code, which means we must
release the runtime system after calling caml_startup as we did in an
earlier commit.

(cherry picked from commit 9d4b87e0338d45456c0506fd2c25b80f7ef6b1e1)