tests: Explicitly add -L../plugins/ocaml/.libs to find -lnbdkitocaml
The actual libnbdkitocaml.so library is located in
plugins/ocaml/.libs, but we were only specifying plugins/ocaml as the
search path.  How did it ever work?  Turns out using LD_LIBRARY_PATH
set in wrapper.c.  It seems unreliable to rely on this, so set the
path properly, as we are already doing elsewhere.