curl: Use curl multi interface
See the comment at the top of plugins/curl/pool.c for general
information about how this works.

This makes a very large difference to performance over the previous
implementation.  Note for the tests below I also applied the next
commit changing the behaviour of the connections parameter.

Using this test case:

  $ http=https://cloud-images.ubuntu.com/lunar/current/lunar-server-cloudimg-amd64.img
  $ nbdkit -r -U - curl $http ipresolve=v4 --run 'nbdcopy -p $uri null'

The times are as follows:

  multi, connections=64           17.2s
  multi, connections=32           21.7s
  wget                            28.4s
  multi, connections=16           41.3s
  before this commit             180s