pool: Add outline get_ready and after_fork functions
In a forthcoming commit we will need to create a multi handle and a
background thread, requiring use of the .get_ready (for multi) and
.after_fork (for the thread) plugin methods.  This commit removes the
empty load_pool function and adds pool_get_ready and pool_after_fork,
and the associated machinery in curl.c.

This commit on its own does nothing, future commits will fill in these
functions with useful work.

Reviewed-by: Eric Blake <eblake@redhat.com>