server: Allow filters to append output to --dump-plugin
There was no way for filters to append their own output to:

  nbdkit <plugin> --filter=<filter> --dump-plugin

This could be useful since some filters (especially qcow2dec) have
configurable elements (like zstd support) which could be queried by
other programs.

I thought about adding a --dump-filter flag, but it would be quite
awkward to implement.  Overloading the existing --dump-plugin flag
makes much more sense, and with proper namespacing of the output
should not cause any issues.

The suggested way to probe for features in a filter will be to use the
null plugin:

  nbdkit null --filter=<filter> --dump-plugin

I did not document this (in nbdkit-probing(1)) in the current commit
because no existing filter implements this method yet.  I will instead
add support for this method to qcow2dec filter and then document it in
a subsequent commit.

This commit documents the new method only (in nbdkit-filter(3)).