ocaml: Always register the thread in every callback
There was an undetected crash in OCaml plugins' .unload, where we
called caml_acquire_runtime_system.  This was only reproducible when
using the nbdkit -s option, which was only used by one test
(test-ocaml-errorcodes).

The crash didn't cause the test to fail since we don't check the exit
code of nbdkit when running it from nbd_connect_command as in this
particular test.

Looking at the OCaml runtime the crash was because for some reason
when .unload was called, we were in an unregistered thread.  OCaml
wants to allocate some thread local storage and requires you to
register each thread in the program.  If you don't do this, a common
symptom is Caml_check_caml_state crashes when it tries to read the
thread local storage, which is what happened here.  The program's main
thread is registered by default.

I am not clear why it was crashing in .unload since (with nbdkit -s)
we ought to be on the main thread (always registered), and other
callbacks which happen on the main thread (eg. .config_complete) also
call caml_acquire_runtime_system without crashing.

I don't fully understand the bug; nevertheless, it is safe and cheap
to always call caml_c_thread_register, even in the main thread (or
when we're supposed to be in the main thread) so let's do that.

(cherry picked from commit 533d0aa18c2b66656cd42c1e3776fe8c0a784b72)
Signed-off-by: Eric Blake <eblake@redhat.com>