ocaml: Properly link plugins depending on OCaml 4 or 5
For OCaml 5, plugins should be linked with unix.cmxa and threads.cmxa.
In addition, in OCaml 5 these libraries are now located in
subdirectories of ocamldir so you must use -I +unix -I +threads.

For OCaml 4, plugins should only be linked with unix.cmxa.  If they
are linked with threads.cmxa then they will hang when forking because
this library adds enforcement of mutexes (ie. implements them using
pthread_mutex) and does not work correctly across fork.  (OCaml 5
always enforces mutexes and fixes the fork bug.)  Some discussion of
this bug here:
https://discuss.ocaml.org/t/fatal-error-fatal-error-during-lock-resource-deadlock-avoided/12457

Deal with this with a configure-time test, and modify how we link
examples and tests, and documentation.

After this change, with OCaml 5 ./configure should print:

  checking major version of OCaml... 5
  checking for what libraries are needed for OCaml plugins... -I +unix unix.cmxa -I +threads threads.cmxa

and for OCaml 4:

  checking major version of OCaml... 4
  checking for what libraries are needed for OCaml plugins... unix.cmxa

On OCaml 4 (only) we must also avoid using
caml_c_thread_{,un}register, since these are defined by threads.cmxa
(or rather, the C library that this pulls in).  This avoids the
following error:

  nbdkit: error: cannot open plugin './test-ocaml-plugin.so': plugins/ocaml/.libs/libnbdkitocaml.so.0: undefined symbol: caml_c_thread_unregister