podwrapper: Add --replace parameter
This allows you to replace literal text in the manual page.  The
common use is to use this like AC_SUBST.  For example if @FOO@ was
defined by ./configure you could use this in a Makefile.am:

  podwrapper.pl ... --replace "__FOO__:$(FOO)"

to replace __FOO__ in the manual page with @FOO@ from a ./configure
AC_SUBST([FOO]) statement.