Revert "tests: Use --exit-with-parent in the test framework"
When compiling tests which use the test framework on macOS, we see
this error:

Undefined symbols for architecture arm64:
  "_nbdkit_debug", referenced from:
      _exit_with_parent_loop in libutils.a(libutils_la-exit-with-parent.o)
  "_nbdkit_error", referenced from:
      _exit_with_parent_loop in libutils.a(libutils_la-exit-with-parent.o)
  "_nbdkit_shutdown", referenced from:
      _exit_with_parent_loop in libutils.a(libutils_la-exit-with-parent.o)
ld: symbol(s) not found for architecture arm64

This happens because the exit-with-parent.c file in libutils.a on
macOS calls various nbdkit_* functions (and needs to, eg it must call
nbdkit_shutdown).  This is fine from the server, but doesn't work when
calling standalone test programs.

Anyway, there is a much better way to test for --exit-with-parent
functionality which I will add in the following commit.

This reverts commit c6299889e80217cfadf488a67961be9eb6d24e38.