protocol: Add definitions for extended headers
We share nbd-protocol.h with libnbd.  That project is taking the lead
on supporting NBD 64-bit extensions (see upstream nbd commits 36abf47d
and a9384e2f on the extension-ext-header branch[1]); it will be a
while longer before nbdkit can do so, because first we need to add a
v3 plugin protocol that lets plugins support 64-bit operations.

But in the short term, it doesn't hurt to add additional protocol
types, even if we aren't using them.

[1] https://github.com/NetworkBlockDevice/nbd/blob/extension-ext-header/doc/

Signed-off-by: Eric Blake <eblake@redhat.com>