cc: Allow configuration without absolute paths
In https://gitlab.com/nbdkit/nbdkit/-/merge_requests/30, Khem reports
that in a cross-compilation environment, nbdkit embeds the absolute
name of the cross-compiler into the resulting cc plugin, even though
running the plugin should probably be favoring the bare name 'cc' or
'gcc'.  This in turn leads to non-reproducible builds.  As the goal of
cross-compiling nbdkit is to produce a binary that behaves identically
regardless of the build environment used, this means we need to give
the user control over the defaults for CC and CFLAGS embedded into the
cc plugin.

However, instead of trying to munge the build environment variable as
suggested in that merge request, I found it cleaner to just add
additional precious variables to be set at configure time, as in:

./configure CC=/path/to/cross-compiler CC_PLUGIN_CC='ccache gcc' ...

Viewing './configure --help' now shows extra lines, such as:

  CC_PLUGIN_CC
              Value to hard-code into the cc plugin's default for CC, instead
              of $CC

Reported-by: Khem Raj
Fixes: https://gitlab.com/nbdkit/nbdkit/-/merge_requests/30
Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <20230818132748.1153048-1-eblake@redhat.com>
[eblake: drop doc changes, reword configure help output]
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>