curl: Rename pool to worker thread
Since commit a74b289ee1 ("curl: Use curl multi interface") it no
longer makes sense to talk about a pool of handles, when in fact there
is only a single multi handle and one easy handle per request (not
pooled or reused in any way).

However there is a worker thread.  So for consistency with the VDDK
plugin which uses the same thread model rename pool to worker thread.

This is mostly just renaming, with a few comment changes.