server: Calculate $uri in one place
Move the calculation of $uri to the main function (well, inlined
there), and out of --run code.

This is largely code motion.  In theory it changes the content of $uri
since we now shell quote it after generating it, but this ought not to
have any practical effect.

Thanks: Eric Blake for code suggestions
Reviewed-by: Eric Blake <eblake@redhat.com>