fuzzing/README: Remove advice to set ASAN_OPTIONS
At least with newer Clang and AFL++ it's both not necessary to set
ASAN_OPTIONS explicitly, and AFL++ sets the environment variable and
chooses a better default.

In particular I found that the existing advice caused leak checking to
be enabled (which is very undesirable for fuzzing as we usually exit
the binary on an error path).  This occurs even though our
recommendation included detect_leaks=false.  That could be because our
syntax was wrong or something has changed upstream.