python: Clarify possible handle values returned by open() method
Previously we stated that any "non-NULL" Python value could be
returned as the handle, which doesn't make any sense.  In fact any
Python object can be returned, even None.  The way to indicate an
error is to throw an exception.  Clarify the documentation and
example.

This goes all the way back to when the Python plugin was introduced in
commit 185539eed1 ("python: New plugin that lets you write plugins as
Python scripts", Feb 2014).

Fixes: commit 185539eed1aafa601ee2ed3471ff8d868d23cba5