pattern: Accurately update b & offset in loop
It's better to update the 'b' pointer and the 'offset' using the
calculated increment 'n', rather than using '8-o' which might be
larger than the remainder of the buffer.  Previously we could create a
dangling pointer beyond the end of the output buffer, and this should
not matter, but it's best not to do this if we can easily avoid it.

Also improve the test by replacing qemu-io with nbdsh.  It is able to
make non-aligned requests so we can now test those.

Reviewed-by: Eric Blake <eblake@redhat.com>