server: Add magic values to important structs
I accidentally passed an incorrect parameter to a next function in a
filter.  I wrote:

  next->can_write (handle);

instead of the correct form:

  next->can_write (next);

Static checking didn't catch this because handle has type void *.
It took me a very long time to work out what the problem was.

In this commit, add some dynamic asserts that we are passing correct
structs around to help find these sorts of bugs in future, and to
better catch any internal inconsistency.

With this commit, the same mistake above results in the assertion:

  nbdkit: threadlocal.c:260: threadlocal_push_context: Assertion `ctx->magic == CONTEXT_MAGIC' failed.

with stack trace:

  #6  0x0000000000420639 in threadlocal_push_context (ctx=0x7f8415bd7540)
      at threadlocal.c:260
  #7  0x0000000000407440 in backend_can_write (c=0x7f8415bd7540) at backend.c:458
  #8  0x00007f841667a3eb in readonly_can_write (next=0xcfca160,
      handle=0x7f8415bd7540) at readonly.c:92

Reviewed-by: Eric Blake <eblake@redhat.com>