server: Small tidy up of globals
Separate globals used for command line parameters from other globals.

socket_activation is not declared extern or used outside server/main.c
so make it static.

Document random_fifo* variables.

Move documentation of 'top' into server/main.c for consistency with
other comments in this file.  (It would probably be better to document
this in server/internal.h but this is the smaller diff.)