partition: Suggest alternate partition-sectorsize
When we reach this error it means that we have failed to detect the
"EFI PART" signature (indicating GPT) and we've fallen back to parsing
MBR, but in doing so we have discovered a GPT protective MBR which
should only happen for GPT.  A possible cause for missing the
signature was because we have the wrong sector size.

Therefore check for the current sector size (which should be either
512 or 4096) and suggest that the user sets the other sector size.

Also avoids the case where the user already set partition-sectorsize=4k
and we were suggesting that they set it again.

Reported-by: Ming Xie
Fixes: commit 7b9301a4c569456a4f96784229a2cd48e8957662
Fixes: https://issues.redhat.com/browse/RHEL-19815
(cherry picked from commit cd761c9bf770b23f678fd82f0d1c8d4cce2ed1b5)