server: Add --no-meta-contexts option
It is nice to be able to configure nbdkit to behave like other common
servers to ease interoperability testing.  With nbd-server 3.24, there
was no support for structured replies, so --no-sr was good enough; but
nbd-server 3.25 learned structured replies for the sake of sparse
reads while still leaving meta contexts unsupported.  The existing
noextents filter is insufficient to emulate this behavior of
nbd-server, since nbdkit has defaulted to supplying its own all-data
"base:allocation" context when the plugin does not support extents.

Note that libnbd 1.15.3 through current 1.18.2 cause
nbd_can_meta_context() to fail with EINVAL when the server supports
structured replies but lacks meta contexts; but back when I made that
change in libnbd, the manual testing I did via a one-off hack to
nbdkit, as documented in libnbd commit 55b09667, did not expose this
corner case that we now see in nbd-server.  While failure was a
desirable change in light of the libnbd API addition of
nbd_opt_set_meta_context() for when the server fails after a partial
list, it was an unintended regression when paired with nbd-server and
the more common automatic handshake process when there is no list at
all.  The test case added here works whether or not that bug in libnbd
has been fixed.

Signed-off-by: Eric Blake <eblake@redhat.com>