common/protocol: Error out if generate-protostrings.sh fails
Previous commit b4ecd758c9 ("configure: Check that bash is
sufficiently new") was added because older bash did not support
'declare -A' (for associative arrays).

A symptom of the failure was that this script failed silently,
generating a corrupted C file which caused a later compilation error.

Use 'set -e' to force an error here instead.