perl: Fix crash if script parameter is missing
$ nbdkit perl
Segmentation fault (core dumped)

The crash happens when we reach config_complete, script was not
defined amongst the config parameters, the Perl interpreter is
therefore not started, and we crash trying to call into Perl to check
if "config_complete" is defined.

Fix by rechecking if script is set in this callback.  After this
change:

$ nbdkit perl
nbdkit: error: the first parameter must be script=/path/to/perl/script.pl

Fixes: commit 6edf6b4e261011342c40e871231d8430db31f490