eval: Check that get_size method is provided
Previously you could start nbdkit-eval-plugin with no parameters:

 $ nbdkit eval

In this state it runs but does nothing useful and always gives an
error when you connect:

  nbdkit: eval[1]: error: /tmp/nbdkitEL7JDh/missing: the get_size method is required

In the sh plugin we aren't able to check if the required method(s) are
present (because that plugin uses an external shell script).  But we
are able check the command line arguments for this plugin.