server: C-quote filenames before printing them
If the filename contains strange or non-printable characters, then
it's better to quote it rather than mess up the output.  We had a
report of a case where DOS line endings were used in a shebang plugin,
like this:

  #!/usr/sbin/nbdkit python\r\n

What happens here is that Linux ran nbdkit with the parameter
"python\r", resulting in the peculiar error message:

  $ /path/to/plugin.py
  -plugin.so: cannot open shared object file: No such file or directory

The complete error message contained several literal '\r' characters,
resulting in the first part of the error being overwritten when
printed.

After this change, the error message becomes:

  $ ./nbdkit /path/to/plugin.py
  nbdkit: error: cannot open plugin "/home/rjones/d/nbdkit/plugins/python\r/.libs/nbdkit-python\r-plugin.so": /home/rjones/d/nbdkit/plugins/python\r/.libs/nbdkit-python\r-plugin.so: cannot open shared object file: No such file or directory

which may be clearer.

Reported-by: Mykola Ivanets
Thanks: Eric Blake